+++ commands/music/pause.ts
@@ -28,6 +28,7 @@ export default class Pause extends Command {
         const message = this.message
         const embeds = new Embeds(discord, message)
         const audio = new Audio(discord, message)
+        if (!audio.checkMusicPlaying()) return
         audio.pause()
         const rep = await message.reply("Paused the song!")
         rep.delete({timeout: 3000}).then(() => message.delete().catch(() => null))
